fixed demo examples
